fix(commands): add top-level context command
