+++ src/types/command/commandBase.ts
@@ -5,7 +5,7 @@ import { unalias } from './unalias'
 
 const commandMap = new Map<string, string[]>([
   ['compile', ['job', 'service', 'identify']],
-  ['context', ['list', 'create', 'edit', 'delete']],
+  ['context', ['create', 'delete', 'edit', 'export', 'list']],
   ['folder', ['create', 'delete', 'list']],
   ['doc', ['init', 'lineage']],
   ['servicepack', ['deploy']]
